Replace using of the f16 by the f32

+++ src/game/Viewport.zig
@@ -18,8 +18,8 @@ cols_pad: u8,
 pub fn init(rows: u8, cols: u8) Viewport {
     return .{
         .region = .{ .top_left = .{ .row = 1, .col = 1 }, .rows = rows, .cols = cols },
-        .rows_pad = @intFromFloat(@as(f16, @floatFromInt(rows)) * 0.2),
-        .cols_pad = @intFromFloat(@as(f16, @floatFromInt(cols)) * 0.2),
+        .rows_pad = @intFromFloat(@as(f32, @floatFromInt(rows)) * 0.2),
+        .cols_pad = @intFromFloat(@as(f32, @floatFromInt(cols)) * 0.2),
     };
 }
